Make mawk use the whole number as array subscript
Because the address numbers of puzzle 2 day 14 can get quite big and
mawk doesn't handle big numbers very well.

My first idea was to convert the numbers to strings by concatening
an empty string. Interestingly, this doesn't work because mawk already
crops the number in the process.

So `sprintf("%.0f", …)` is the only way I know of to convert the whole
number to a string on mawk.