Simplify solution for day 15
Everything becomes so much simpler when `i`, the round counter,
is 0 based. Then, it's a simple matter of saving when the number
was spoken last.

Also, use an environment variable to set the number of games and
use part1 in part2 because they only differ by the number of
games.